start to make org config pluggable
